Correctifies permissions for data management

+++ app/Http/Controllers/ActivityController.php
@@ -23,21 +23,27 @@ class ActivityController extends Controller
      */
     public function index()
     {
-        $showForm = false;
+        if (\Auth::user()->is_admin) {
 
-        // TODO: It would be really cool if this worked.
-        // if ($errors->any()) {
-        //     $showForm = true;
-        // }
+            $showForm = false;
 
-        $activities = \App\Activity::orderBy('name')->get();
+            // TODO: It would be really cool if this worked.
+            // if ($errors->any()) {
+            //     $showForm = true;
+            // }
 
-        foreach ($activities as $activity) {
-            $activity->numberOfGroups = $activity->groups()->count();
-            $activity->numberOfParticipants = $activity->participants();
+            $activities = \App\Activity::orderBy('name')->get();
+
+            foreach ($activities as $activity) {
+                $activity->numberOfGroups = $activity->groups()->count();
+                $activity->numberOfParticipants = $activity->participants();
+            }
+
+            return view('activities.index', compact('activities', 'showForm'));
         }
 
-        return view('activities.index', compact('activities', 'showForm'));
+        return redirect()->route('home');
+
     }
 
     /**
@@ -47,7 +53,7 @@ class ActivityController extends Controller
      */
     public function create()
     {
-        //
+        return redirect('/activities');
     }
 
     /**
@@ -59,17 +65,23 @@ class ActivityController extends Controller
     public function store(Request $request)
     {
 
-        $validatedData = $request->validate([
-            'activityName' => 'required|unique:activities,name',
-            'activityDescription' => 'required',
-        ]);
+        if (\Auth::user()->is_admin) {
 
-        $activity = new \App\Activity;
-        $activity->name = $request->input('activityName');
-        $activity->description = $request->input('activityDescription');
-        $activity->save();
-        $request->session()->flash('status', 'Activity created!');
-        return redirect()->route('activities.index');
+            $validatedData = $request->validate([
+                'activityName' => 'required|unique:activities,name',
+                'activityDescription' => 'required',
+            ]);
+
+            $activity = new \App\Activity;
+            $activity->name = $request->input('activityName');
+            $activity->description = $request->input('activityDescription');
+            $activity->save();
+            $request->session()->flash('status', 'Activity created!');
+            return redirect()->route('activities.index');
+
+        }
+
+        return redirect()->route('home');
     }
 
     /**
@@ -79,8 +91,8 @@ class ActivityController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show($id)
-    {
-        //
+    {  
+        return redirect('/activities');
     }
 
     /**
@@ -91,8 +103,15 @@ class ActivityController extends Controller
      */
     public function edit($id)
     {
-        $activity = \App\Activity::find($id);
-        return view('activities.edit', compact('activity'));
+        if (\Auth::user()->is_admin) {
+
+            $activity = \App\Activity::find($id);
+            return view('activities.edit', compact('activity'));
+
+        }
+
+        return redirect()->route('home');
+
     }
 
     /**
@@ -105,17 +124,24 @@ class ActivityController extends Controller
     public function update(Request $request, $id)
     {
 
-        $validatedData = $request->validate([
-            'activityName' => 'required|unique:activities,name,' . $id,
-            'activityDescription' => 'required',
-        ]);
+        if (\Auth::user()->is_admin) {
 
-        $activity = \App\Activity::find($id);
-        $activity->name = $request->input('activityName');
-        $activity->description = $request->input('activityDescription');
-        $activity->save();
-        $request->session()->flash('status', 'Activity updated!');
-        return redirect()->route('activities.index');
+            $validatedData = $request->validate([
+                'activityName' => 'required|unique:activities,name,' . $id,
+                'activityDescription' => 'required',
+            ]);
+
+            $activity = \App\Activity::find($id);
+            $activity->name = $request->input('activityName');
+            $activity->description = $request->input('activityDescription');
+            $activity->save();
+            $request->session()->flash('status', 'Activity updated!');
+            return redirect()->route('activities.index');
+        
+        }
+
+        return redirect()->route('home');
+
     }
 
     /**
@@ -126,8 +152,16 @@ class ActivityController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        \App\Activity::destroy($id);
-        $request->session()->flash('status', 'Activity deleted!');
-        return redirect()->route('activities.index');
+
+        if (\Auth::user()->is_admin) {
+
+            \App\Activity::destroy($id);
+            $request->session()->flash('status', 'Activity deleted!');
+            return redirect()->route('activities.index');
+
+        }
+
+        return redirect()->route('home');
+
     }
 }
